Project restructure. This restructure will allow us to more easily share code (like logging package). Added basic dockerfile for api
