--- core/README.md
@@ -1,1 +1,0 @@
-# Server Health Monitor Core
