--- core/go.mod
@@ -1,3 +1,0 @@
-module github.com/PR-Developers/server-health-monitor/core
-
-go 1.16
