--- data-collector/.gitignore
@@ -1,2 +1,0 @@
-vendor/
-data-collector.log
