--- data-collector/README.md
@@ -1,1 +1,0 @@
-# Server Health Monitor Data Collector
