--- data-collector/go.mod
@@ -1,5 +1,0 @@
-module github.com/PR-Developers/server-health-monitor/data-collector
-
-go 1.16
-
-require github.com/shirou/gopsutil/v3 v3.21.6
