+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/PR-Developers/server-health-monitor
+
+go 1.16
+
+require github.com/shirou/gopsutil/v3 v3.21.6
